Changed type of various times from float to double
git-svn-id: http://svn.code.sf.net/p/petavision/code/PVSystemTests@6095 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ InitWeightsTest/InitWeightTestProbe.hpp
@@ -17,7 +17,7 @@ public:
 	InitWeightTestProbe(const char * filename, HyPerLayer * layer, const char * msg);
 	InitWeightTestProbe(HyPerLayer * layer, const char * msg);
 
-	virtual int outputState(float timef);
+	virtual int outputState(double timef);
 
 };
